docker/taist_bot_api/requirements.txt
fastapi
uvicorn
line-bot-sdk
requests
jinja2
python-dotenv